♻️ [Refactoring]: #455 Phase 7 line_continuation カテゴリを tests/line_continuation.rs に分離
行継続（バックスラッシュ + EOL）テスト 3 件を
literal_string/tests/line_continuation.rs に切り出し、tests.rs ハブに
`mod line_continuation;` 宣言を追加。
